all products of a specific user delete api integtate and all products get api update with products of a specific user

+++ src/components/UserProfile.jsx
@@ -5,23 +5,34 @@ function UserProfile() {
 
     const navigate = useNavigate();
     const auth = localStorage.getItem("user");
+    const userId = JSON.parse(localStorage.getItem("user"))._id;
 
     const handleLogout = () => {
         localStorage.clear("user");
         navigate("/login");
     };
+
     const handleDelete = async (id) => {
-        const api = `http://localhost:5000/account/${id}`; 
-        const createRequest = {
-            method: "Delete",
-        };
-        let result = await fetch(api, createRequest);
-        await result.json();
-        if (result) {
-            alert("Your account will permanently delete from database!!")
-            localStorage.clear("user");
-            navigate("/signup");
-        };
+        const confirmation = window.confirm("Are you sure you want to delete your account and all data? This action cannot be undone.");
+        if (confirmation) {
+            const api = `http://localhost:5000/account/${id}`;
+            const createRequest = {
+                method: "Delete",
+            };
+            let result = await fetch(api, createRequest);
+            result = await result.json();
+        
+            if (result) {
+                const api2 = `http://localhost:5000/products-of-user/${userId}`;   // also delete products of specific user.
+                const createRequest2 = {
+                    method: "Delete",
+                };
+                await fetch(api2, createRequest2);
+
+                localStorage.clear("user");
+                navigate("/signup");
+            }
+        }
     };
 
     return (
@@ -31,7 +42,7 @@ function UserProfile() {
                 <h3 className="user-profile-card__name">{JSON.parse(auth).name}</h3>
                 <p className="user-profile-card__email">{JSON.parse(auth).email}</p>
                 <button className="user-profile-card__logout" onClick={handleLogout}>Logout</button>
-                <button className="user-profile-card__logout __delete" onClick={()=>handleDelete(JSON.parse(auth)._id)}>Account delete</button>
+                <button className="user-profile-card__logout __delete" onClick={() => handleDelete(JSON.parse(auth)._id)}>Account delete</button>
             </div>
         </>
     );
